+++ NEWS.md
@@ -8,6 +8,7 @@
 * Improve consistency of "season" variable, now always an ordered factor.
 * `trakt.api.call` doesn't throw an error but only a warning on trakt availability
   * Idea is to not completely fail if trakt throws a 5xx error
+* Fix up `trakt.stats` since it's now functional.
 
 ### v0.13.0